+++ src/shared/enums/TableTypeEnum.ts
@@ -1,0 +1,4 @@
+export enum TableTypeEnum {
+  NODES = 'nodes',
+  COUNTRIES = 'countries'
+}
